Coding - Scope 31 Graphic3d enums as enum class
Converts 31 unscoped Graphic3d enums to enum class (out of 43
candidates). Value names are preserved; callers updated to scoped
form. Localized static_cast<int> added where enum class values
interact with int-typed APIs (array subscripts, anonymous _NB
counters, dump macros, switch/case).

Kept as plain enum because of pervasive int interoperation or
bitflag semantics:
- Graphic3d_CappingFlags, Graphic3d_TransModeFlags,
  Graphic3d_ShaderFlags, Graphic3d_TextureSetBits,
  Graphic3d_DiagnosticInfo, Graphic3d_TypeOfShaderObject
  (all bitflag)
- Graphic3d_TextureUnit, Graphic3d_DisplayPriority (integer
  indices/priorities compared against int, used as array
  subscripts throughout the rendering pipeline)
- Graphic3d_TypeOfShadingModel (has obsolete aliases and is
  integer-valued with V3d_* aliases)
- Graphic3d_TypeOfAttribute, Graphic3d_TypeOfData,
  Graphic3d_TypeOfBackground (passed as int / GLint to GL APIs)
- Graphic3d_StereoMode (array subscript and int comparison)
- Graphic3d_TypeOfLightSource (array subscript + int cast in
  counters)
- Graphic3d_TypeOfTextureMode (compared with GL_NONE int)
- Graphic3d_NameOfMaterial (stored as int in TDF framework)

+++ src/Visualization/TKService/Graphic3d/Graphic3d_DisplayPriority.hxx
@@ -38,7 +38,8 @@ enum Graphic3d_DisplayPriority
 enum
 {
   Graphic3d_DisplayPriority_NB =
-    Graphic3d_DisplayPriority_Topmost - Graphic3d_DisplayPriority_Bottom + 1
+    static_cast<int>(Graphic3d_DisplayPriority_Topmost)
+    - static_cast<int>(Graphic3d_DisplayPriority_Bottom) + 1
 };
 
 #endif // _Graphic3d_DisplayPriority_HeaderFile
