+++ src/Visualization/TKService/Graphic3d/Graphic3d_FrameStatsData.hxx
@@ -48,13 +48,13 @@ public:
   double ImmediateFrameRateCpu() const { return myFpsCpuImmediate; }
 
   //! Get counter value.
-  size_t CounterValue(Graphic3d_FrameStatsCounter theIndex) const { return myCounters[theIndex]; }
+  size_t CounterValue(Graphic3d_FrameStatsCounter theIndex) const { return myCounters[static_cast<int>(theIndex)]; }
 
   //! Get counter value.
   size_t operator[](Graphic3d_FrameStatsCounter theIndex) const { return CounterValue(theIndex); }
 
   //! Get timer value.
-  double TimerValue(Graphic3d_FrameStatsTimer theIndex) const { return myTimers[theIndex]; }
+  double TimerValue(Graphic3d_FrameStatsTimer theIndex) const { return myTimers[static_cast<int>(theIndex)]; }
 
   //! Get timer value.
   double operator[](Graphic3d_FrameStatsTimer theIndex) const { return TimerValue(theIndex); }
@@ -123,16 +123,19 @@ public:
   double& ChangeImmediateFrameRateCpu() { return myFpsCpuImmediate; }
 
   //! Return a timer object for time measurements.
-  OSD_Timer& ChangeTimer(Graphic3d_FrameStatsTimer theTimer) { return myOsdTimers[theTimer]; }
+  OSD_Timer& ChangeTimer(Graphic3d_FrameStatsTimer theTimer)
+  {
+    return myOsdTimers[static_cast<int>(theTimer)];
+  }
 
   //! Get counter value.
-  size_t& ChangeCounterValue(Graphic3d_FrameStatsCounter theIndex) { return myCounters[theIndex]; }
+  size_t& ChangeCounterValue(Graphic3d_FrameStatsCounter theIndex) { return myCounters[static_cast<int>(theIndex)]; }
 
   //! Modify counter value.
   size_t& operator[](Graphic3d_FrameStatsCounter theIndex) { return ChangeCounterValue(theIndex); }
 
   //! Modify timer value.
-  double& ChangeTimerValue(Graphic3d_FrameStatsTimer theIndex) { return myTimers[theIndex]; }
+  double& ChangeTimerValue(Graphic3d_FrameStatsTimer theIndex) { return myTimers[static_cast<int>(theIndex)]; }
 
   //! Modify timer value.
   double& operator[](Graphic3d_FrameStatsTimer theIndex) { return ChangeTimerValue(theIndex); }
