+++ src/Visualization/TKService/Graphic3d/Graphic3d_TypeOfBackground.hxx
@@ -28,7 +28,8 @@ enum Graphic3d_TypeOfBackground
 
 enum
 {
-  Graphic3d_TypeOfBackground_NB = Graphic3d_TOB_CUBEMAP + 1
+  Graphic3d_TypeOfBackground_NB =
+    static_cast<int>(Graphic3d_TOB_CUBEMAP) + 1
 };
 
 #endif // _Graphic3d_TypeOfBackground_HeaderFile
